chore: Refactor packages to a simpler hierarchy (#7)
Originally, the resources and data sources were in their own package
(organized by GitHub API endpoint types, e.g. `repositories`). Having a
set hierarchy and multiple packages has created complexity and import
loops when trying to introduce tests to the provider.

In order to simplify this, resources and data sources are now part of
the `provider` package. Keeping `functions` separate is a design
decision, for now.