Fix double apply bug
This commit fixes a bug when updating the name of a repository using
Terraform, an error is displayed but the name updates on the GitHub side
with no issue.

Then on a second run, the Terraform apply runs successfully (no changes)
and the state correctly reflects the change.

This is not ideal for users since they get an error and are unsure if
their change was successful (it was) and is a bad experience.

By pulling in both the plan and state data, the correct values are
"trued" up.